asn1c: fix IMPORTS of same name from multiple modules
ASN.1 supports importing the same name from different modules
as well as allowing that name to be in the current module -
refer to ITU-T X.680 and https://stackoverflow.com/a/67910959.

Rework the IMPORTS tracking in _imp_ from dict name -> module
dict name -> list(module, ...), and resolve types that way.

Fixes #56